Optimized utilities

+++ utility/unfollow.js
@@ -1,22 +1,9 @@
 import firebase from '../lib/firebase'
 
-const unfollow = async (authId, userId) => {
+const unfollow = (authId, userId) => {
     const firestore = firebase.firestore()
-    await firestore
-        .collection(`users/${authId}/followings`)
-        .doc(userId)
-        .delete()
-        .catch((error) => {
-            return false
-        })
-    await firestore
-        .collection(`users/${userId}/followers`)
-        .doc(authId)
-        .delete()
-        .catch((error) => {
-            return false
-        })
-    return true
+    firestore.collection(`users/${authId}/followings`).doc(userId).delete()
+    firestore.collection(`users/${userId}/followers`).doc(authId).delete()
 }
 
 export default unfollow
